refactor: Renamed folders to be singular instead of plural
