+++ backend/src/main/java/com/homeorholiday/backend/controller/UserController.java
@@ -1,7 +1,7 @@
-package com.homeorholiday.backend.controllers;
+package com.homeorholiday.backend.controller;
 
 import com.homeorholiday.backend.dto.CreateUserInput;
-import com.homeorholiday.backend.models.User;
+import com.homeorholiday.backend.model.User;
 import com.homeorholiday.backend.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.graphql.data.method.annotation.Argument;
