+++ backend/src/main/java/com/homeorholiday/backend/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.homeorholiday.backend.repository;
 
-import com.homeorholiday.backend.models.User;
+import com.homeorholiday.backend.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
